requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0

PyGithub==2.4.0
GitPython==3.1.43
python-dotenv==1.0.1

PyJWT==2.9.0
cryptography==43.0.3
